requirements-api.txt
# MNIST inference API (serve_mnist_api.py) and Render deployment
-r requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
Pillow>=9.0.0
